Archival Commit of what the code would be for initalization in Servo constructor declaration
THIS IS FOR ARCHIVAL PURPOSES. The previous code had the illusion of working during my test because I only reset the esp32 after upload which does not depower it. Had I done so, the PCA9685 would have lost power and as such needed to be re initialized, which was broken.

 During the normal servo obj creation portion of the firmware, I2C is not actually initialized yet, so a dummy objforsetupServo needs to be created after that step. Well either that, or I2C would need to be initialized separately from the servo interface, but with the PCA being the only thing to use I2C that doesn't make a ton of sense. There also isn't a nice place to insert it without modifying _main